SFIR-162 Move importLandData to its own file for code/test reuse
