+++ src/tests/import-tests/agreements.import.test.js
@@ -6,7 +6,7 @@ import {
   deleteFiles
 } from '~/src/tests/import-tests/setup/s3-test-helpers.js'
 
-import { importLandData } from '~/src/features/land-data-ingest/workers/ingest.module.js'
+import { importLandData } from '~/src/features/land-data-ingest/workers/import-land-data.js'
 import { connectToTestDatabase } from '~/src/tests/db-tests/setup/postgres.js'
 import { getRecordsByQuery } from '~/src/tests/import-tests/setup/db-helper.js'
